refactor: move modal component
